fix off-by-one business days counting
workdays' networkdays() method includes the start date in the total
number of days if it is a business day.

In other words, we could get "1" when the start date and end date are
the same day. If today's a business day, and the release is tomorrow, we
will get "2", etc.

Subtract one to make this easier to reason about, because start_date
will almost always be a business day. (We'll always query the bot during
work).