task: fix returning None on missing product
Prior to this change, if we found no product value in ReleaseTask's
from_text(), we would pass "None" to ReleaseTask's canonical_product(),
which then bailed out trying to parse the None value with a regex.

Test for None earlier so we don't end up passing None into the
canonical_* methods.